chore(about): move about data into data/about (#46)
