--- apps/data/palestine_data.py
@@ -1,32 +1,0 @@
-class PalestineData:
-    palestine = {
-        "title": "STAND WITH PALESTINE 🇵🇸",
-        "description": "Taking Action for <span class=\"text-transparent bg-clip-text bg-gradient-to-r from-green-600 to-red-600\">Humanity in Gaza</span>",
-        "callToAction": "This is a call to action. The humanitarian crisis in Gaza—marked by displacement, hunger, and loss—demands our compassion and urgent response.",
-        "statistics": {
-            "url": "https://app.powerbi.com/view?r=eyJrIjoiODAxNTYzMDYtMjQ3YS00OTMzLTkxMWQtOTU1NWEwMzE5NTMwIiwidCI6ImY2MTBjMGI3LWJkMjQtNGIzOS04MTBiLTNkYzI4MGFmYjU5MCIsImMiOjh9",
-            "peopleKilled": {
-                "headline": "51,000+",
-                "details": "People killed since October 2023",
-                "source": "OCHA",
-                "url": "https://www.ochaopt.org/content/reported-impact-snapshot-gaza-strip-15-april-2025"
-            },
-            "peopleDisplaced": {
-                "headline": "1.9M+",
-                "details": "People displaced—90% of Gaza's population",
-                "source": "UNRWA",
-                "url": "https://www.unrwa.org/resources/reports/unrwa-situation-report-167-situation-gaza-strip-and-west-bank-including-east-jerusalem"
-            },
-            "peopleInjured": {
-                "headline": "116,343+",
-                "details": "People injured since October 2023",
-                "source": "OCHA",
-                "url": "https://www.ochaopt.org/content/reported-impact-snapshot-gaza-strip-15-april-2025"
-            }
-        },
-        "actions": {
-            "donateToUNRWA": "https://www.unrwa.org/donate",
-            "supportMedicalAid": "https://donate.who.foundation/"
-        },
-        "whyIStand": "As someone who values <em class=\"italic\">justice</em> and <em class=\"italic\">humanity</em>, I'm moved by the suffering in Gaza. The crisis in Gaza isn't just a headline—it's a call to support dignity and hope for millions. Join me in making a difference.",
-    }
